backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pillow>=10.0
speciesnet
gunicorn>=21.0
